Add new match modal, update routing, and clean up unused code
