+++ src/routes/scouting/index.tsx
@@ -1,17 +1,9 @@
-import { Autocomplete, NumberInput } from "@/components/inputs";
-import { type Match, matchCollection, newId, useReactivity } from "@/data/db";
-import {
-  ActionIcon,
-  Button,
-  LoadingOverlay,
-  Modal,
-  Table,
-  isNumberLike,
-} from "@mantine/core";
-import { isNotEmpty, useForm } from "@mantine/form";
+import { DeleteModal, NewMatchModal } from "@/components/modals";
+import { matchCollection, useReactivity } from "@/data/db";
+import { ActionIcon, Button, Table } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
 import { IconPencil, IconTrash } from "@tabler/icons-react";
-import { createFileRoute, useNavigate } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import { useMemo, useState } from "react";
 
 export const Route = createFileRoute("/scouting/")({
@@ -66,8 +58,10 @@ function Page(): JSX.Element {
       />
       <DeleteModal
         opened={deleteModalOpened}
-        matchId={matchId}
         onClose={closeDeleteModal}
+        onDelete={() => {
+          matchCollection.removeOne({ id: matchId! });
+        }}
       />
 
       <div className="flex my-10">
@@ -146,136 +140,5 @@ function ActionGroup({
         <IconTrash />
       </ActionIcon>
     </div>
-  );
-}
-
-function NewMatchModal({
-  opened,
-  onClose,
-  scouters,
-  teams,
-}: {
-  opened: boolean;
-  onClose: () => void;
-  scouters: string[];
-  teams: number[];
-}): JSX.Element {
-  const navigate = useNavigate();
-  const [loading, setLoading] = useState(false);
-
-  const form = useForm<Omit<Match, "started" | "finished" | "phases">>({
-    mode: "uncontrolled",
-    validate: {
-      scouter: isNotEmpty("Scouter name is required"),
-      matchNumber: (value) =>
-        value <= 0 || value > 999 || !isNumberLike(value)
-          ? "Invalid match number"
-          : null,
-      teamNumber: (value) =>
-        value <= 0 || value > 9999 || !isNumberLike(value)
-          ? "Invalid team number"
-          : null,
-    },
-  });
-
-  const handleSubmit = async (values: typeof form.values) => {
-    setLoading(true);
-
-    const id = newId();
-    matchCollection.insert({
-      id,
-      ...values,
-      started: new Date(),
-      phases: {},
-    });
-
-    await new Promise((resolve) => setTimeout(resolve, 500));
-    setLoading(false);
-
-    onClose();
-    navigate({
-      to: "/scouting/$matchId/collect/pre-match",
-      params: { matchId: id },
-    });
-  };
-
-  return (
-    <Modal
-      opened={opened}
-      onClose={onClose}
-      title="New Match"
-      centered
-      overlayProps={{
-        backgroundOpacity: 0.55,
-        blur: 3,
-      }}
-    >
-      <LoadingOverlay visible={loading} />
-
-      <form
-        className="flex flex-col gap-y-2"
-        onSubmit={form.onSubmit(handleSubmit, (errors) => {
-          const firstErrorPath = Object.keys(errors)[0];
-          form.getInputNode(firstErrorPath)?.focus();
-        })}
-      >
-        <Autocomplete
-          label="Scouter"
-          data={scouters}
-          {...form.getInputProps("scouter")}
-        />
-        <Autocomplete
-          label="Team"
-          data={teams.map((t) => t.toString())}
-          inputMode="numeric"
-          {...form.getInputProps("teamNumber")}
-        />
-        <NumberInput label="Match" {...form.getInputProps("matchNumber")} />
-        {/* <NumberInput label="Team" {...form.getInputProps("teamNumber")} /> */}
-        <Button type="submit" className="ml-auto mt-2">
-          Go
-        </Button>
-      </form>
-    </Modal>
-  );
-}
-
-function DeleteModal({
-  opened,
-  matchId,
-  onClose,
-}: {
-  opened: boolean;
-  matchId: string | null;
-  onClose: () => void;
-}): JSX.Element {
-  const handleDelete = () => {
-    if (matchId) {
-      matchCollection.removeOne({ id: matchId });
-    }
-    onClose();
-  };
-
-  return (
-    <Modal
-      opened={opened}
-      onClose={onClose}
-      title="Delete Match"
-      centered
-      overlayProps={{
-        backgroundOpacity: 0.55,
-        blur: 3,
-      }}
-    >
-      <div className="flex flex-col gap-y-4">
-        <div>Are you sure you want to delete this match?</div>
-        <div className="space-x-4 ml-auto">
-          <Button color="red" variant="subtle" onClick={handleDelete}>
-            Delete
-          </Button>
-          <Button onClick={onClose}>Cancel</Button>
-        </div>
-      </div>
-    </Modal>
   );
 }
